bpf: Allow uneven numbers in `memcpy` and `memcmp`
The `memcpy` and `memcmp` builtin replacements did work for uneven
values other than 1. This commit updates both builtin replacements so
they do. This is quite important since it isn't uncommon to copy strings
with an uneven amount of chars in the new test framework.

Additionally, this commit extends the max size of values passed into
`memcmp` from 32 bytes to 72 bytes which is required for some unit tests
to compare packets which are larger than 32 bytes.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>