bgp: BGP Control Plane modularization
This work converts the BGP Control Plane controller and BGP Route
Manager into hive cells, leaving as much of the existing code intact.
These cells are now hooked into the agent hive directly. The daemon
now takes the Controller as parameter both to preserve the behavior of
setting the controller as a field value on the daemon and so the BGP
controllers lifecycle events are invoked.

Follow up commits can break the package up into more discrete parts to
aid in testing the individual components and or mocking them out.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>