docs: Corrected bad config example in BGP docs.
The original example advised the user to use a `NotIn` label selector
with an empty array. But in practice this isn't allowed resulting in
the following error: `[]string(nil): for 'in', 'notin' operators,
values set can't be empty`.

So recommend the user to use a dummy key and value instead of just a
dummy key.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>