resources,lbipam: WIP resource gossip for more stable tests
NOTE: This is a work in progress.

This commit adds an additional option for `resource.Resource[T].Events`
which allows the subscriber to attach a "gossip" channel which will
relay messages about normally internal state. Specifically when events
are consumed and when they are marked done. The use case for this
gossip channel is to use it in module/cell tests to determine when all
events, provided to a cell via resources have been processed. This
replaces the time based approach which is subject to flaking if the
tests are ran on a congested or slow host machine.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>