Remove `HAVE_SOCKET_LOOKUP`
The value of `HAVE_SOCKET_LOOKUP` was dependent on the result of a
feature probe that tests for the presence of the `bpf_skb_change_tail`
helper function, which was added in kernel v4.9. Now that the minimum
supported kernel version is v5.4, we can remove the probe and assume
we always have this feature available.

Given the existence of global asserts for features significantly newer,
I think its safe to not add an explicit assert for this feature.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>