node/manager: Validate node implementation on running device changes
We currently have logic in the node manager which periodically
revalidates the node implementations of all running nodes. This is good
but in certain situations such as when a carrier goes down, it can take
a while before this validation takes place.

This commit adds logic to the node manager background sync to watch the
devices table for changes to the set of running devices, physical
devices and to reconcile all node implementations when a change is
detected.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>