datapath/neighbor: Simplify netlink handle creation
We were using the `netns.Current` to get the "current" namespace and
then passing the FD to `netlink.NewHandleAt`. This is unnecessary as
`netlink.NewHandle` already creates a handle in the current namespace
by default.

The original code had a race where the netns handle could be finalized
before the netlink handle was created.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>