pkg/loadbalancer/maps: use map spec registry
Update pkg/loadbalancer/maps to use the new map spec registry as the
source of map specifications. The cilium-dbg logic is updated to no
longer use the generated specs, it now just opens the pinned maps.
The compile time config for the LB maps are now removed. And the
skiplb cell has been reworked to use the map spec registry.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>